LayoutTests/http/tests/security/xss-DENIED-invalid-domain-change-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/iframe-invalid-domain-change.html from frame with URL http://127.0.0.1:8000/security/xss-DENIED-invalid-domain-change.html. Domains, protocols and ports must match.


PASS: cross-site not access allowed
